Add resources readiness checks for helm/manifest
This commit adds Readiness checks for workload resource types based
on the resource status provided by Kubernetes.

It fixes a bug for the helm and manifests deployers that occurs when
trying to update resources with immutable field that are set on the
server side like ClusterIP or the ResourceVersion.

It fixes a bug for the helm and manifests deployers that
could occur during the deletion of the deploy items if the managed
resources in status have never been updated due to errors in the
Kubernetes manifests.

For the helm deployer, it removes the decoding of the status
when deleting the files because it is already done during the
creation of the new Helm struct.

Signed-off-by: lcavajani <lcavajani@suse.com>